Remove circular dependencies
