+++ Handler/IntegerOptions.ts
@@ -1,0 +1,4 @@
+export interface IntegerOptions {
+	min?: number
+	max?: number
+}
