+++ Handler/PriceOptions.ts
@@ -1,0 +1,6 @@
+import { isoly } from "isoly"
+
+export interface PriceOptions {
+	currency?: isoly.Currency
+	toInteger?: boolean
+}
