+++ Handler/price.ts
@@ -5,11 +5,7 @@ import { Settings } from "../Settings"
 import { State } from "../State"
 import { StateEditor } from "../StateEditor"
 import { add } from "./base"
-
-export interface PriceOptions {
-	currency?: isoly.Currency
-	toInteger?: boolean
-}
+import { PriceOptions } from "./PriceOptions"
 
 class Handler implements Converter<number>, Formatter {
 	readonly currency: isoly.Currency | undefined
